+++ ui/ui_dasm.h
@@ -83,8 +83,8 @@ typedef struct {
     int num_layers;
     const char* layers[UI_DASM_MAX_LAYERS];
     void* user_data;
-    int init_x, init_y;
-    int init_w, init_h;
+    float init_x, init_y;
+    float init_w, init_h;
     bool open;
     bool valid;
     uint16_t start_addr;
@@ -128,10 +128,10 @@ void ui_dasm_init(ui_dasm_t* win, const ui_dasm_desc_t* desc) {
     win->read_cb = desc->read_cb;
     win->start_addr = desc->start_addr;
     win->user_data = desc->user_data;
-    win->init_x = desc->x;
-    win->init_y = desc->y;
-    win->init_w = desc->w;
-    win->init_h = desc->h;
+    win->init_x = (float) desc->x;
+    win->init_y = (float) desc->y;
+    win->init_w = (float) desc->w;
+    win->init_h = (float) desc->h;
     win->open = desc->open;
     win->highlight_color = 0xFF30FF30;
     for (int i = 0; i < UI_DASM_MAX_LAYERS; i++) {
